Add remaing methods in handlers.rb
