+++ main.rb
@@ -1,10 +1,4 @@
 require_relative 'handlers'
-require_relative 'person'
-require_relative 'student'
-require_relative 'book'
-require_relative 'classroom'
-require_relative 'teacher'
-require_relative 'rental'
 
 class App
   include Handlers
@@ -15,7 +9,7 @@ class App
   end
 
   def start
-    puts 'Welcome to School Library! 😀'
+    puts 'Welcome to School Library!😀'
     loop do
       menu
       option = gets.chomp
@@ -23,7 +17,7 @@ class App
 
       get_num option
     end
-    puts 'Thank you for using our Library!'
+    puts 'Thank you for using our Library!😀'
   end
 
   def menu
@@ -55,71 +49,6 @@ class App
     else
       puts 'Please select a number between 1 and 7. ❌'
     end
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-
-    print 'Name: '
-    name = gets.chomp
-
-    print 'Specialization: '
-    specialization = gets.chomp
-
-    teacher = Teacher.new(age, specialization, name)
-    @people << teacher
-
-    puts 'Teacher created successfully'
-    sleep 0.75
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-
-    print 'Author: '
-    author = gets.chomp
-
-    book = Book.new(title, author)
-    @books << book
-
-    puts 'Book added successfully'
-    sleep 0.75
-  end
-end
-
-def create_rental
-  puts 'Select a book from the following list by number'
-  @books.each_with_index { |book, index| puts "#{index}) Title: #{book.title}, Author: #{book.author}" }
-
-  book_id = gets.chomp.to_i
-
-  puts 'Select a person from the following list by number (not id)'
-  @people.each_with_index do |person, index|
-    puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-  end
-
-  person_id = gets.chomp.to_i
-
-  print 'Date: '
-  date = gets.chomp.to_s
-
-  rental = Rental.new(date, @books[book_id], @people[person_id])
-  @rentals << rental
-
-  puts 'Rental created successfully'
-  sleep 0.75
-end
-
-def list_rentals_by_person_id
-  print 'ID of person: '
-  id = gets.chomp.to_i
-
-  puts 'Rentals:'
-  @rentals.each do |rental|
-    puts "Date: #{rental.date}, Book '#{rental.book.title}' by #{rental.book.author}" if rental.person.id == id
-    sleep 0.75
   end
 end
